Rewrite a lanedo patch to fix a bug
We should not be calling init on the object provided by the [CALayer layer]
static factory method as the layer is already initialised.

We also *must* call retain on it as it's an autoreleased object which we are
storing indefinitely in a local var.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=8472