Patch mcs.in to setup PKG_CONFIG_PATH on Mac.
Removes pkg-config from the list of binaries we symlink into /usr/bin.

The Mono root is appended in $PATH so that the Mono-installed pkg-config
will be available as a fallback. If you have a copy from Homebrew, it
will use that one instead.

The Mono root is prepended in $PKG_CONFIG_PATH so that
Mono-installed .pc files are looked up first.