Update mono config generation
Mono PR mono/mono#1064 changes how the paths for libMonoPosixHelper
and libgdiplus stored in the configuration file are computed. This
means that no fixing is needed anymore for libMonoPosixHelper, because
the intallation prefix is automatically added. Instead libgdiplus is
assumed to be in the default library load path unless another path is
specified at configure-time through the --with-libgdiplus option.