[package] fix building with artifacts from a previous run
When running bockbuild without removing artifacts from
previous runs (i.e. to save bandwidth by not downloading
tarballs again, or to save CPU to avoid compiling again),
many 'File not found' errors were happening, such as
'../../autoconf.py not found'.

The reason was that ../../autoconf.py is a relative path,
and as such, it's very fragile that the access to it gets
broken by placing calls to os.chdir() everywhere in the
code.

The solution is to:
a) use self.pushd(foo) instead of self.cd(foo), so that we can
b) use self.popd() after we have finished working with 'foo'