[package] fix regression about patch downloading
There are some packages which contain patches in the sources
which need to be retrieved before being applied. Recent
refactoring introduced this bogus call to self.rm() which
would remove the workspace of a package before trying to
download the second source (the workspace is at this point
very important because we have already used tar to extract
what we downloaded in the first source).