[solitary] Inspect assemblies in a temp path (bgo#751339)
The call to Assembly.GetTypes() requires that all the dependencies
of a certain assembly be in the same folder as where it is, otherwise
a ReflectionTypeLoadException could be thrown (this starts happening
somehow only since upgrading mono 2.x to 4.0.1).

This was happening when trying to bundle banshee, because for example
its Banshee.GStreamerSharp.dll assembly depended on Banshee.Services.dll
which was placed in a different folder.

The way to fix this is to copy all .NET assemblies in a first pass to
a temporary directory, where they can be inspected via reflection later
(this is why we replace IEnumerable<X> with List<X>, because we must
ensure a first pass is completed, not yielding, before we start calling
into AssemblyItem::Load() on all instances).

Fixes part of https://bugzilla.gnome.org/show_bug.cgi?id=751339