Change the way workspaces are made  for “local git repos” (e.g. The master repo driving the build).
Instead of applying patches and building directly, treat the local repo as a local Git cache, which is the standard behavior for the rest of the bockbuild-managed packages.
Also use the new shadow copying for creating the second workspace needed when building 32/64
This method should be more robust, and the shadow copying should reduce disk space use.