Fix a long-standing tar/xz bug.
xz's configure script would find gettext in the staging directory if it hasn't been nuked from last time. When it finally gets nuked, xz stops working, missing a libintl.8.dylib.

Better solutions would be to add the stage directory in the $PATH after the toolchain build, and building a separate gettext for the toolchain (not possible right now)